Turn Plugins module into global plugins container
This is now less code and it's consistent with components registry
